sf patch #1348112, from Bj�rn Beskow This should fix the problems with merging (bug #1348109).
I also removed some code that attempted to merge the 'name' field of PackageData and SourceFileData.  The name is used as the key in the hash map when finding the objects to merge together, so the two names will ALWAYS be the same.

That's something we could add an "assert" for, if we were using assertions.


git-svn-id: https://cobertura.svn.sourceforge.net/svnroot/cobertura/trunk@317 3373f243-8d0f-0410-a8fe-cd033897b1b5