The rest of sf patch #1446503, from John Lewis
1. Archives within archives are instrumented, but this
is only done if <includeClassname regex=""> has been
specified.

3. If jars are instrumented, remove the signature files.

6. The code that checks to see if a class matches with
any <includeClasses regex=""> element is now aware of
the WEB-INF/classes dir that may be in an archive. The
WEB-INF/classes prefix is removed before a match is
checked.

7. When specifying a todir, instrumented archives are
placed in a location that preserves its relative
position to the fileset's basedir instead of just being
put in the todir. In other words, if a fileset
includes a jar that is a few directories down (ex.
a/b/c/d.jar), then the instrumented jar will be placed
in todir/a/b/c/d.jar instead of todir/d.jar. It is
assumed that if the current behavior is desired, a
simple <copy> will do the trick.


git-svn-id: https://cobertura.svn.sourceforge.net/svnroot/cobertura/trunk@404 3373f243-8d0f-0410-a8fe-cd033897b1b5