When generating Javadoc, only try to generate documentation for the .java files.  This was apparently problematic when trying to generate Javadoc using the GNU javadoc tool 'gjdoc'
This was pointed out to me by David Walluck, who sent me this email:
"Hi Mark.

I just built cobertura 1.9, and I ran into a problem
building the javadocs with gjdoc.

In turns out that without a fileset, it will try to import
and parse everything as a source file (including package.html).

This is the simple fix I applied:"
(the diff follows)


git-svn-id: https://cobertura.svn.sourceforge.net/svnroot/cobertura/trunk@528 3373f243-8d0f-0410-a8fe-cd033897b1b5